+++ platform-sdk/consensus-event-stream/src/timingSensitive/java/org/hiero/consensus/event/stream/StreamingPerformanceTest.java
@@ -3,7 +3,7 @@ package org.hiero.consensus.event.stream;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import com.swirlds.common.utility.StopWatch;
+import com.swirlds.base.time.StopWatch;
 import java.util.Iterator;
 import java.util.LinkedList;
 import java.util.List;
