+++ platform-sdk/swirlds-base/src/main/java/com/swirlds/base/time/StopWatch.java
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.common.utility;
+package com.swirlds.base.time;
 
-import com.swirlds.base.time.Time;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.Objects;
 import java.util.concurrent.TimeUnit;
